backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
yfinance==0.2.18
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
requests==2.31.0
slowapi==0.1.9
redis==5.0.1
aioredis==2.0.1
celery==5.3.4
faker==20.1.0
